jimt/CT-446 - checkForDuplicate* functions as utilities, not hooks
