+++ apps/src/codebridge/FileBrowser/hooks/index.ts
@@ -1,3 +1,1 @@
-export * from './useCheckForDuplicateFilename';
-export * from './useCheckForDuplicateFoldername';
 export * from './useHandleFileUpload';
